Refactor Description View
