+++ CoreLocationComponent/CoreLocationComponent/Localizable.xcstrings
@@ -168,6 +168,9 @@
         }
       }
     },
+    "Undefined" : {
+
+    },
     "We haven't found the optimal time yet" : {
       "localizations" : {
         "id" : {
