+++ CoreLocationComponent/CoreLocationComponent/Model/DescriptionModel.swift
@@ -1,0 +1,14 @@
+//
+//  DescriptionModel.swift
+//  CoreLocationComponent
+//
+//  Created by Nathanael Juan Gauthama on 16/07/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct DescriptionModel {
+    var conclusionDescription: LocalizedStringKey
+    var timeDescription: LocalizedStringKey
+}
